Module Declaration: Thread through most internal Hack systems
Summary:
# High Level
This stack threads module declarations through `hh_server` such that they are now understood and are accessible across files. The main effect is that declaring a file to be a part of a module requires that said module has an existing declaration in the current codebase. For example:
```
//// a-decl.php
<?hh
<<file:__EnableUnstableFeatures('modules')>
module A {}

//// a.php
<?hh

// Unbound name error w/o `a-decl.php`
<<file:__EnableUnstableFeatures('modules'), __Module('A')>>
```
# This Diff

This adds *the vast majority* of the plumbing required to reach our end state of modules stored as decls, s.t. their information can be accessed across files.

## `ClientIdeIncremental`

Based on the name of the module, my assumption is that this is meant for file updates *within* the IDE when the changes haven't been saved.

**This is currently not implemented, but is vital to implement for a good user experience in VScode.** This looks like it's implemented via the facts parser, which AFAIK doesn't handle modules. Any changes related to adding / removing module decls will not be observed until saved.

As per Lucian:
> what is ClientIdeIncremental? - the IDE typechecks the unsaved-files-in-the-editor against files on disk for the rest of WWW. If a file on disk changes, then watchman notifies us about it, and tells VSCode, and VSCode sends us a didChangedWatchedFile, and we invoke ClientIdeIncremental. In this way the IDE will know to typecheck its editor buffers against the new updated rest-of-the-codebase.
Without implementing this, then if the user has a file open which says "error you claim to be part of module A but module A isn't declared", and the user modifies a file on disk so that module A is now declared, then the IDE will be unable to know that fact without restarting.

Lucian has also noted that this entire logic could be implemented via direct decl parsing, so this could become a non-issue (as we already have implemented modules in direct decl parsing).

## Decl Diffing

Modules can be placed on classes, functions, and types. Functions and types appear to need no adjustment here as we diff them via `Poly.( = )`. Classes I've made internal-ness and module membership into "big diff" changes. This *requires* potentially re-checking all uses of the class.

## Decl Heap

Module decls add a new heap kind, `Modules`. This looks *roughly* like the heap for constants, rather than classes, as we're making modules *case sensitive*. I didn't store this as part of an existing heap as modules are really their own *thing:* they're a separate construct that we can't really categorize among the existing Hack *things*. ljw1004 wrote out a good visualization:
```
1. Fun
2. GConst
3. Typelike
   3.1. Typedef
   3.3. Classlike
        3.3.1. Class
        3.3.2. Trait
        3.3.3. Enum
        3.3.4. Interface
4. Modules
```
Modules are obviously unlike constants or functions, and (for now) cannot be referenced in expression positions, making them unlike types.

**Modules are case-sensitive,** unlike functions / classes.

## Dependencies

I've added a new kind of dependency, a `Dep.Module`: this is not yet in use, but this represents data flow from a module to its members. One can imagine the following code:
```
<?hh

<<file: __Module('A')>>

function foobar(): void {}
```
... having an edge in the depgraph:
```
Module A  -> Fun foobar
```

## HackC (`DeclProvider`)

For Decls-In-Compilation, because modules don't exist at runtime, it doesn't make any sense to try to expose modules at the moment. If you ask for a module name, we will tell you that it doesn't exist, unconditionally.

## Naming

As noted previously, modules live *along side* other constructs in Hack, thus we create:
* A new entry in the naming table (`Naming_sqlite`)
* A new kind of naming heap (`Naming_heap.Modules`)

We thread namespace elaboration through any attached attributes as well, as part of naming.

## Providers

Similar to the other heaps, we've threaded through a new kind of *thing,* copy-pasting the logic from constants, with modules where needed. The code in `hphp/hack/src/providers/` is boilerplate-y, wouldn't be surprised if there's one copy-paste error kicking about.

## Server Functionality (`hphp/hack/src/server/`)

`ServerExtractStandalone` ignores module decls for now. Realistically, we should include them as it could be useful for repro-ing module specific bugs.

`ServerLazyInit` bit me: there's a part where we remove references to any potentially changed decls when initializing from saved state that I forgot to fill out for modules.

Differential Revision: D34481050

fbshipit-source-id: 26f6de8cc2e1d9ebdcf13087c62714bea4f2e0d5