Module Declaration: Add Syntax
Summary:
# High Level
This stack threads module declarations through `hh_server` such that they are now understood and are accessible across files. The main effect is that declaring a file to be a part of a module requires that said module has an existing declaration in the current codebase. For example:
```
//// a-decl.php
<?hh
<<file:__EnableUnstableFeatures('modules')>
module A {}

//// a.php
<?hh

// Unbound name error w/o `a-decl.php`
<<file:__EnableUnstableFeatures('modules'), __Module('A')>>
```

# This Diff

This diff adds / modifies data structures that will be used later in the stack for module declaration.
* Adds the `md_annotation` field to the existing `module_def` node. This is used down the line in `typing_toplevel` as a place to store TAST info for a given module decl.
* Adds `module_def_type`, which will be used to store module info across files

Differential Revision: D34480788

fbshipit-source-id: 2490d6e5052cc72fb8bbabd948fe002ae181e723