Only set module *once* in direct decl mode
Summary:
The way we determine module membership in direct decl mode is by hooking into `make_file_attribute_specification`; the file attributes themselves are dropped, but we take this moment to scan them for the `__Module` attribute.

Unfortunately, the following code is currently legal:
```
<?hh

<<file:__Module('A')>>
<<file:__Module('B')>>

function f(): void {}
```

Legacy decls claim that `f` belongs to module `A`, so let's just be consistent here: we set the module of the included decls according to the *first* `__Module` attribute we see.

Reviewed By: jamesjwu

Differential Revision: D34408893

fbshipit-source-id: b835d521ebde083346feb5a5f6319b40851722c5