check StackLimit in rust_parser_errors
Summary:
It turns out there is another recursive traversal of the AAST after lowering,
for error checking. And it's now the new largest peak stack user (in hphp/test).

We were not measuring it at all, but without StackLimit it causes a stack overflow.
This plumbs in a StackLimit, checks it, and reports it's peak separately from the
other peak values.

Reviewed By: shayne-fletcher

Differential Revision: D34931121

fbshipit-source-id: ec341fd4c417ab7b9047fcc000904a153f07e1e1