Disallow internal class members in non-internal traits
Summary:
For the simplicity of the module feature, we will require that traits with `__Internal` members must themselves be `__Internal`. Consider the following:
```
//// A.php
<<file:__Module('A')>>
trait Foo {
  <<__Internal>> public int $x = 0;
}
```
```
//// B.php
class Bar {
  use Foo;
  public function getX(): int {
    return $this->x;
  }
}
```
Whether or not this should error depends on the semantics of traits:
* If traits are reified constructs in Hack, akin to classes, then this should error: clearly `Bar` is not in module `A`, and `Foo::$x` is internal to module `A`
* If traits are "copy-paste" then this shouldn't error: the body of `Foo` is no longer in a module.

Instead of trying to answer this question, we disallow methods and properties in non-internal traits from themselves being internal.

Reviewed By: jamesjwu

Differential Revision: D34742245

fbshipit-source-id: 0f042d55a5b115c699232711dc7fbde4af0d85a1