stream_get_meta_data should not crash the HHVM server if called on a closed file
Summary:
`HPHP::File::getMetadata()` calls `HPHP::File::eof()`:

https://www.internalfb.com/code/fbsource/[5135cddb168dc8e0968e931c8416fb6dbd88638c]/fbcode/hphp/runtime/base/file.cpp?lines=450-463

Which uses an assert to verify that the given file or stream hasn't been closed:

https://www.internalfb.com/code/fbsource/[14db0a0e6c0bd92511ad7aa392c6a04fb733183f]/fbcode/hphp/runtime/base/plain-file.cpp?lines=248-255

This means that a user can write userspace Hack code which crashes HHVM servers, or at least the HHVM servers that are running with assertions enabled. That's bad.

Reviewed By: ricklavoie

Differential Revision: D34706612

fbshipit-source-id: 587f1d0f6e4459bd74946907da5ed52746076fbe