Add framework for running code *after* an opcode is executed
Summary:
This makes certain things much easier - and in particular, is needed
for implementing closures, since the `ObjectData*` used to identify a closure
is only created *after* the interpreter is run (we have no way of knowing it
when we process the opcode ourselves). This will also make it possible to more
precisely track collections in the future.

This gets used in the next diff - putting this up separately so we can bikeshed
the design.

I put this in `iopPreamble` because we know that will *always* run no matter
the opcode, so it's the easiest place to put it. I also decided to use an enum
to identify what to run, as we can dispatch to whatever additional
functionality is needed based on that. Storing lambdas and the like would
probably be too much overhead given we need this rarely.

Reviewed By: dkgi

Differential Revision: D34946155

fbshipit-source-id: 3ea788886c754179ec4534307867a53e65f3cadc