Add support for functors
Summary:
This was half the work needed to implement closures, so split it out
into its own commit for easier review.

Figuring out where to put the parameters on the stack is a bit confusing since
we have to predict *how* hhvm will adjust the stack (since we run before it does).
I did some trial/error and testing across various scenarios after reading the
code and this should work in the majority of the cases we care about.

Note for any future people who need to touch this code: logging the stack
addresses here, and then logging them again when the *next* opcode runs were
immensely helpful.

Reviewed By: dkgi

Differential Revision: D34946157

fbshipit-source-id: 6f39db930044355683107402eb7353469ed92014