Clarify desired input state of resume helpers and audit their usage
Summary:
There was a weird distinction between resumeHelper* and interpHelper* that the
former accepted interpreter-like context, while the latter TC-like context.

In some cases this worked just by an accident because of an extra VM reg sync.
However, func entries will soon have a significant distinction and it won't be
possible to satisfy both contexts at the same time.

Create both an interp and a TC version for each of them and audit their usage.

In general, *FromInterp stubs should never be used in the same context as
*FromTC stubs or translation addresses. This is still being violated by the
handleResume() function itself and will be addressed by the next diff.

Reviewed By: ricklavoie

Differential Revision: D34886903

fbshipit-source-id: e6edaec5a88dd65a8195ff86baffddb9490b06fb