Redo +/-/* to fix arithmetic typehole and result types with like-type arguments
Summary:
There were two problems in type checking +/-/*, both of which contributed to allowing the following to type check.
```
function expect_int(int $n) : void {
}

function f<T as num>(T $n) : void {
  $m = Vector {};
  $m[] = $n;
  expect_int($m[0] + 1);
}

<<__EntryPoint>>
function main() : void {
  f(1.1);
}
```

- The case logic assumed if the arguments to + didn't definitely include a `float` or a `num`, it was enough to check if 1 was an `int`, but that is unsound if one argument is an `int` and one inconclusive, e.g., an unsolved variable.

- When used with a default, `expand_type_and_narrow` had an special case to allow use of the default, even if it was not a super type of the type being narrowed. This is not sound, and it allowed narrowing the unsolved type variable to `int` w/o constraining it.

This diff removes the unsound case from `expand_type_and_narrow` and adds some extra parameters to allow finer grained control over what is being early solved. These control whether it should force early solving when it has nothing to widen to, and whether it should allow `nothing` as a result. The former is necessary to try to solve to `float`, but not force the solving if there is no float lower bound. The latter is necessary to handle like-types.

Like types were not handled uniformly, and cases regarding `dynamic` were spread throughout the code, and there were some cases that did not give the correct result. Now they are handled in analogous way to the rule for calling a `supportdyn` function: if the argument has a like type, it is removed, and the result gets a like type. This separates the like-type logic from the overloading logic for the binop.

The only non-obvious (unary) case is where we have:
`-$d : t` where `$d : dynamic`. We could reasonably say that `t = dynamic`, using the logic above, since `dynamic = ~nothing`. That didn't fit into the code well, since `expand_type_and_narrow` doesn't want to force solve to `nothing`, so we add an optional argument to allow it to solve to nothing. In the binary case, we return `dynamic` when both arguments were dynamic, but otherwise return a like type of the other argument.

Reviewed By: andrewjkennedy

Differential Revision: D35053012

fbshipit-source-id: 2ea6625a2c37c4083073e474867f2679e7447535