Move PERDIR ini settings to be SYSTEM settings
Summary:
These PERDIR settings are not configurable with `ini_set()` yet have a per
request allocated callback.  The underlying storage is only configured at
system initialization time (unless things overwrite it, which would be racy if
it is happening in a request since these are all backed by global storage).

An alternative to changing the mode of these settings would be for PERDIR
settings to be equivalent to system settings in configuration.  I am not sure
how much we care matching the setting modes of PHP.

While here I delete EnableZendIniCompat since we have no ambition of being
compatible with zend.  Despite the current default being true, I opted to
maintain more of the false behavior since it seems more appropriate.

Reviewed By: ricklavoie

Differential Revision: D35060722

fbshipit-source-id: 8df8693e01a578cbebe9e1f6b453be7bff63fadb