Avoid point-insert into flat_map in find_alias_sets
Summary:
We currently do a point-insert into a flat_map, which can be expensive.

https://www.boost.org/doc/libs/1_61_0/doc/html/boost/container/flat_set.html#idp16460480-bb
> Complexity: Logarithmic search time plus linear insertion to the elements with bigger keys than x.

I believe the incoming may_alias sets are empty, and we do indeed insert in ascending order, so that probably helps.  But maybe we can do better?  Here I construct a (trivially) ordered/uniqued vector<ASetID> and use adopt_sequence to std::move it into the flat_set:

https://www.boost.org/doc/libs/1_65_1/doc/html/boost/container/flat_map.html#idp47782160-bb
> Complexity: Assuming O(1) move assignment, O(1)

Reviewed By: ricklavoie

Differential Revision: D35071898

fbshipit-source-id: 479e370086fa8ba942b1049060398017177f13a7