Store a bespoke decl_error type on decl_class_type instead of Errors.t
Summary: Errors.t is a collection of *rendered* error messages with a complex representation, which makes producing it on the Rust side difficult: we'd have to copy some error-rendering logic. Since the possible error values which can occur on a folded class are a very small set, encode that small set as a new type, making FFI much easier.

Reviewed By: shayne-fletcher

Differential Revision: D34913558

fbshipit-source-id: 1b956530c78b4b99aab9f37d8f80dd5cf4c20677