Fix memory leak for text output in HPHPc
Summary:
If you specify format=text, we'll use Unit::toString() to
dump the units. Unit::toString() calls Func::prettyPrint(), which uses
Func::getLineNumber() to provide line numbers. However
Func::getLineNumber() relies on the treadmill to free old
memory. We're not in request context here in HPHPc, so that memory
will never get freed (until the process exits) and we leak like a
sieve. For any non-trivial program this quickly OOMs the machine.

Fix it by checking if we're in request context in
Func::prettyPrint(). If we're not, don't print line numbers.

format=hhas doesn't have this problem, as the disassembler uses
different logic.

Reviewed By: paulbiss

Differential Revision: D35071810

fbshipit-source-id: 77aea2633b23248f34a47dc24cc2d1a18c09a97b