Mark create/unwrap_opaque_value as pure
Summary:
Currently, enum class labels and coeffects are unsound in that certain usage of enum class labels can result in coeffect violations (resulting in runtime exceptions). This is a result of enum class labels ultimately being converted into calls to `create_opaque_value` and `unwrap_opaque_value`, which are SystemLib function that are not marked as Pure.

Hack does not check System Lib functions so it didn't report a coeffect type error. And the SystemLib functions are not marked a pure, seemingly as an oversight.

Previously, we would throw a coeffect violation

Reviewed By: oulgen

Differential Revision: D35066883

fbshipit-source-id: f239215bb7e0eb18f7fb51dcce5da5f8bcf5e33f