Add support for supplemental helpers when resuming JIT
Summary:
Change the handleResume() and iop*() APIs to return JitResumeAddr instead of
TCA. The JitResumeAddr allows the handler to be a supplemental helper, that
may adjust the state and continue execution at the real handler passed as
an argument.

This solves the problem that after resumeHelper(), we may want to resume code
in various contexts (start of a translation, return into a translation after
a call) and these contexts require different logic to initialize the TC state
(just sync rvmfp()/rvmsp() regs, also load return regs).

Before this diff we covered all possible situations by simply always loading
the return registers. This approach won't be possible with the upcoming func
entry implementation. This diff prepares the APIs so that the func entry logic
can be plugged in and demonstrates the functionality by cleaning up the return
logic.

The call threaded interpreter needs further updates, tracked by T115315816.

Reviewed By: ricklavoie

Differential Revision: D35136339

fbshipit-source-id: 1245e5403b4f5acca8440c7526848811c0fcc466