refactor for judicious use of for_each
Summary: i appreciate that using `for` over `for_each` is  generally more idiomatic but in this case i argue that its effect of reducing syntactic "noise" makes it worth it? i also argue that after D35063497 it is just more in character with the rest of the program. lastly, ~~even after having to silence clippy~~, there are ~~15~~ 16 less lines to "grok" :)

Reviewed By: edwinsmith

Differential Revision: D35114108

fbshipit-source-id: d890edd17f5337f40da6295ea4f548ecd5e05509