Fix handling of noreturn builtins
Summary:
"noreturn" and "nothing" return types for builtins wasn't
fully implemented, meaning we thought they were objects. So, when we
call the builtin, we expect an object to be returned in the same
location as where the C++ builtin expects the first parameter. As a
result, the first parameter is garbage.

Reviewed By: oulgen, alexeyt

Differential Revision: D35153200

fbshipit-source-id: 1981139f4e4a68d0564ff8ba73241759437acf8d