Remove --target code from HPHPc
Summary:
The only meaningful value for --target is "hhbc" (that's the
only reason we run HPHPc anymore). Remove code for dealing with
it. The option itself remains because I need to remove everything that
passes it beforehand. Also remove the "exe" option for --format. We
haven't supported making executables for quite a while. The other
possible settings are still useful, so keep those.

Reviewed By: oulgen, alexeyt

Differential Revision: D35137455

fbshipit-source-id: 44b523bed4da5349d537213ff4b4495d22e10118