Add a special case for calling like-typed functions
Summary:
When calling a like-typed function `f:~(t1 -> t2)` as `f(e)`, we previoiusly checked `e` twice (because of the Tunion case in the `call` function in `typing.ml`), once with expected type `t1` and once with expected type `dynamic`. This led to some trouble in inferring types in `e` where it contained e.g., a lambda. This could be helpful where `dynamic` is a helpful expected type (e.g., in the `like_higher_order.bad.php` test), but harmful where you need to infer `t1`, as in the `like_call.good.php` test. In general, we'd prefer to not check the argument expression twice, and not accept examples where the programmer could not add a working type annotation because both checks require different annotations.

This change checks `f` as though it were `t1 -> ~t2`, with additional checks that the type of `e` is a dynamic aware sub-type of dynamic. This is similar to the rule for calling `supportdyn` functions, except there we like-strip the argument types, which is not sound to do with the union with dynamic, since the argument type has to be a subtype of `t1` in any case. We parameterise the calling code with which sort of dynamic type we are under to control whether likes are stripped off or not.

This diff also reveals a weakness in typing enum class labels wrt dynamic. See the enum_class_names.bad.php test, which should pass, and did before because of special code in the call function for functions of type `dynamic` which is now bypassed, since we no longer type check like-typed functions following the union with dynamic.

Reviewed By: andrewjkennedy

Differential Revision: D35141550

fbshipit-source-id: a04cc944f8507eab21a35baa34f8bbe486789ea4