Avoid capture when instantiating generic function types
Summary:
Instantiation of function types in `Decl_instantiate.instantiate` fails to avoid "capture" when instantiating under a generic function type. (See https://en.wikipedia.org/wiki/Lambda_calculus#Capture-avoiding_substitutions for the classic definition in the lambda calculus.) This opens up a bad type hole, in which we can launder arbitrary types.

Here is an example (also a test):
```
abstract class Base<T> {
  abstract public function cast<TF>(T $x):TF;
}

// This is bad because we've captured TF
abstract class Bad<TF> extends Base<TF> {
}

class Bad2 extends Bad<mixed> {
  public function cast<TF>(TF $x):TF {
    return $x;
  }
}

function breakit(Base<mixed> $x):int {
  return $x->cast("A");
}
<<__EntryPoint>>
function main():void {
  $b = new Bad2();
  breakit($b);
}
```
To compute the decl for `cast` in `Bad`, we apply the substitution `T := TF` to the function type `(function<TF>(T):TF)`. Naive application of a substitution, which ignores the "clash" with the generic method parameter, produces a type `(function<TF>(TF):TF)`. Obviously this is wrong, as can be seen by alpha-renaming the signature to the equivalent `(function<TF#0>(T):TF#0)` and then applying the substitution to get (correctly) `(function<TF#0>(TF):TF#0)`.

The problem can be avoided if all bound type parameters are unique, e.g. by qualifying them with the class/member name. (This is the approach that C# takes, I believe). Alternatively, one could use some normalized representation such as de Bruijn. Either of these approaches is rather invasive and would require a major overhaul of the type checker.

The easiest approach is to do exactly what is described above: first rename the generic method parameters "away from" all other generics. We only need to do this for names that are present in the target of the substitution, as in the example above.

The renaming of method generics is dependent on a correct treatment of generic method override; currently, we do not permit a method to override one in which the generic parameters have been renamed. That will require a separate diff.

Reviewed By: madgen

Differential Revision: D35022615

fbshipit-source-id: dfba44d6f5e8f015c46cdfa376aabf553ad9f57c