Do not reject lambdas passed to parameters with generic parameter type
Summary: If we attempt to pass a lambda with untyped parameters to a function `foo<T>(T $_)` or  `foo<T as dynamic>(~T $_)` then we reject it as not sufficiently annotated to infer. But in many cases we *can* infer types. Probably this dates from the introduction of new_inference and we were being overly cautious.

Differential Revision: D35212567

fbshipit-source-id: 298286c4408ca1684cd431a33b01b1b7637988b6