Fix expected type for arguments to supportdyn functions
Summary: The arguments to a supportdyn function (when we have failed to type check it using the given type w/o dynamic) are allowed to have like types which are stripped off when compared to the parameters. Update the expected types that are used in computing the types of those arguments to allow like types too.

Reviewed By: andrewjkennedy

Differential Revision: D35215315

fbshipit-source-id: 9290bec26ee33e848d302945bb1059631f2db9fc