unify Local Named/Unnamed variants (2/2)
Summary:
Now that Local::Named and Unnamed are both identified by index,
we don't need to distinguish them at all except in bytecode_printer,
which can do it by `idx`.

Delete Local (the enum) and rename LocalId to Local, and merge
APIs. This harmlessly generalizes LocalRange as a side effect.

Reviewed By: shayne-fletcher

Differential Revision: D35174190

fbshipit-source-id: e807c9e8db255354820cdc19fbed585a15d9d972