Use LocalId for Named locals (1/2)
Summary:
The problem:

HHBC uses numbers (LocalId) for every local, both named and unnamed.
emitter::Local currently uses Str for named locals; they flow through HHAS and aren't
assigned numbers until `as.cpp` converts to HHBC. if we do nothing in hackc,
then HakcCTranslator will need to duplicate the string->id numbering of as.cpp

The solution:

This diff changes `Local::Named(Str)` to `Local::Named(LocalId)` and performs name->id
numbering at emit-time, based on `decl_vars`. Bytecode_printer then uses localid to look
up the string name of a named local, so the hhas output is the same as before.
Since the emitter already figures out which names will be needed and in what order
(in decl_vars), we can use this in the emitter to obtain LocalIds for named locals.

Implications:

We now require the set of named locals to be known in advance before starting
to emit code. If that were not already the case, named and unnamed local ids could
overlap, possibly leading to mis-compiled code. The risk of mis-compilation is serious
enough to warrant compile-time asserts that we correctly declare all named locals.

This diff does not yet erase the distinction between named and unnamed variables.
As a side effect, emitter::Local no longer contains a Str<'arena>, so the 'arena
lifetime param gets dropped from a bunch of code and a number of arena
allocations go away.

Reviewed By: shayne-fletcher

Differential Revision: D34390333

fbshipit-source-id: c2d61f6e76a13062c750d78df64fb6bbf958caa1