Print out HHBC Op enum, not signed integer, from within gdb
Summary:
When trying to view an `Op` enum value in gdb, when the value is positive (e.g. op `SetL`, i.e. 114, since the MSB is 0 in those 8 bits), it currently prints the enum identifier fine.

However, a problem comes when the MSB=1, since what is an unsigned value is currently treated as signed. GDB currently just prints the negative number, rather than treat it as unsigned and print the associated enum identifier. This fixes that.

Reviewed By: oulgen

Differential Revision: D35265984

fbshipit-source-id: 4c990a4fadb2f8dd545bf7eba0bd1345d88197a3