Avoid adding two holes to value collection literals
Summary: ValueCollection literals are special cased in typing since we know that all elements must be subtypes of arraykey. However, since a `Hole` doesn't change the type of an expression, typing will later identify another type mismatch and add a second `Hole`. This diff adds a flag to the helper function `compute_exprs_and_supertype` to avoid doing this

Reviewed By: andrewjkennedy

Differential Revision: D35245209

fbshipit-source-id: df5cdeb58175016766a49e507433530412323a10