Don't print anything when OCaml and rupro decls match
Summary: When the rupro comparator detects no differences, don't print any output. This allows us to run the comparator over a large number of test cases without adding thousands of new expected-output files, which will be useful for enabling the comparator for our typecheck tests. It also helps call attention to the test cases which are still not matching.

Differential Revision: D35320824

fbshipit-source-id: 6a6abfdf81042a1c00446f335e8e9d7c8b400c7f