Make SDT static methods supportdyn
Summary: Static methods that support dynamic type were being intersected with `dynamic` rather than tagged with `supportdyn` which is what top-level functions get. They should be treated similarly. This regresses one test where the intersection was necessary, but lets us use the like-type function rule, which is likely to be more important in practice, given that is the case for top-level functions.

Reviewed By: andrewjkennedy

Differential Revision: D35316385

fbshipit-source-id: b74dc0b924b3f0444d0edddcfdddea26e9fd196c