build naming table on remote workers
Summary:
This diff updates hulk lite to have remote workers build the naming table. This reduces the cost of serializing the naming table to sqlite and also eliminates the cost of uploading/downloading the naming table. To get this working end to end I also had to make a few other changes.

1) Start remote hulk workers as soon as possible. Previously we were doing it very late in the process.
2) Fixed some logic with regards to how we sleep if things aren't done by the time the local worker is done.
3) Hoisted remote worker code to be completely decoupled from the janky XDB/Everstore dependencies. Now we simply start the remote worker -> build naming table -> typecheck -> upload -> die

Reviewed By: nt591

Differential Revision: D35320640

fbshipit-source-id: 8f01b22eeb1eda193acea8760ecab866167a5c35