Reimplement pipe expression logic for readonly inference
Summary:
The pipe expression logic for readonly was getting really messy and also incorrect, so I decided to rewrite it from first principles.

Upon seeing a pipe expression, we recurse on the LHS value. Then, we update the context with $$'s value based on the LHS's readonlyness. Afterward, we recurse on the RHS value.

I make this change only in the typechecker first, and then update the runtime in the diff above this.

Reviewed By: oulgen

Differential Revision: D35301212

fbshipit-source-id: 521e7e2596e61b1622d486abaf400175ef4eb9aa