Do not keep around temporary tyvars longer than necessary
Summary:
Two problems were converging to cause a timeout in our pessimised run of www.

- A list(...) destructure expression creates a list of type vars as buckets for the elements being destructed
  - e.g. list($x) = $vec_int does `vec<int> <: list(#1)`  => `int <: #1`
- The pessimised_builtins flag changes array assignment to create a type var as a bucket for the non-dynamic portion of the element being assigned
  - e.g. a `vec<int>` being assigned a ~int creates #1, does `~int <: ~#1`, and returns `vec<(#1 | int)>`

These type variables were created in the current scope, so they remain unsolved until we close type vars at the end of the function body. The result was that these unsolved type vars were hanging around in the array type (not simplifying the union), and a repeated array assignment was creating an ever-growing union for the array value. These vars are really just temporary buckets and should be solved immediately to whatever value we're putting in them.

This diff makes it so we put a new env on the tyvars stack and then closes it once we've populated the buckets.

Reviewed By: andrewjkennedy

Differential Revision: D35323222

fbshipit-source-id: d1d7ad83952e56a9bd9db1de112a8ab8c784ed58